+++ libs/back/prisma/src/migrations/20241203180946_add_bsvhu_identification_type_enum/migration.sql
@@ -1,0 +1,2 @@
+-- AlterEnum
+ALTER TYPE "BsvhuIdentificationType" ADD VALUE 'NUMERO_IMMATRICULATION';